+++ src/lib/PnP.Framework/AuthenticationManager.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.Identity.Client;
 using Microsoft.Identity.Client.Extensibility;
 using Microsoft.SharePoint.Client;
+using PnP.Core.Services;
 using PnP.Framework.Utilities;
 using PnP.Framework.Utilities.Context;
 using System;
@@ -91,6 +92,7 @@ namespace PnP.Framework
         private readonly ACSTokenGenerator acsTokenGenerator;
         private IMsalHttpClientFactory httpClientFactory;
         private readonly SecureString accessToken;
+        private readonly IAuthenticationProvider authenticationProvider;
         internal CookieContainer CookieContainer { get; set; }
 
         private IMsalHttpClientFactory HttpClientFactory
@@ -237,6 +239,15 @@ namespace PnP.Framework
             return new AuthenticationManager(clientId, clientSecret, userAssertion, tenantId, azureEnvironment, tokenCacheCallback);
         }
 
+        /// <summary>
+        /// Creates a new instance of the Authentication Manager to acquire an authenticated ClientContext.
+        /// </summary>
+        /// <param name="authenticationProvider">PnP Core SDK authentication provider that will deliver the access token</param>
+        /// <returns></returns>
+        public static AuthenticationManager CreateWithPnPCoreSdk(IAuthenticationProvider authenticationProvider)
+        {
+            return new AuthenticationManager(authenticationProvider);
+        }
         #endregion
 
         #region Construction
@@ -541,6 +552,16 @@ namespace PnP.Framework
             // register tokencache if callback provided
             tokenCacheCallback?.Invoke(confidentialClientApplication.UserTokenCache);
             authenticationType = ClientContextType.AzureOnBehalfOf;
+        }
+
+        /// <summary>
+        /// Creates an AuthenticationManager for the given PnP Core SDK <see cref="IAuthenticationProvider"/>.
+        /// </summary>
+        /// <param name="authenticationProvider">PnP Core SDK <see cref="IAuthenticationProvider"/></param>
+        public AuthenticationManager(IAuthenticationProvider authenticationProvider)
+        {
+            this.authenticationProvider = authenticationProvider;
+            authenticationType = ClientContextType.PnPCoreSdk;
         }
         #endregion
 
@@ -900,6 +921,28 @@ namespace PnP.Framework
                         ClientContextSettings clientContextSettings = new ClientContextSettings()
                         {
                             Type = ClientContextType.AccessToken,
+                            SiteUrl = siteUrl,
+                            AuthenticationManager = this
+                        };
+                        context.AddContextSettings(clientContextSettings);
+
+                        return context;
+                    }
+                case ClientContextType.PnPCoreSdk:
+                    {
+                        if (authenticationProvider == null)
+                        {
+                            throw new ArgumentException($"{nameof(GetContextAsync)}() called without an IAuthenticationProvider.");
+                        }
+
+                        var context = GetAccessTokenContext(siteUrl, (site) =>
+                        {
+                            return authenticationProvider.GetAccessTokenAsync(new Uri(site)).GetAwaiter().GetResult();
+                        });
+
+                        ClientContextSettings clientContextSettings = new ClientContextSettings()
+                        {
+                            Type = ClientContextType.PnPCoreSdk,
                             SiteUrl = siteUrl,
                             AuthenticationManager = this
                         };