+++ src/lib/CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- Support for creating a CSOM ClientContext from a PnP Core SDK context [jansenbe - Bert Jansen]
 - Minimal support for TEAMCHANNEL Template #268 [czullu - Christian Zuellig]
 - PnPSDK Mocking + few more cases isolated #262 [mgwojciech - Marcin Wojciechowski]
 - Added implementation of SPWebRequestExecutor that utilizes HttpClient #261 [patrikhellgren - Patrik Hellgren]